Renamed "aux" to "auxil" in folder names and dependencies
