+++ analysis/auxil/__init__.py
