+++ analysis/auxil/aux.py
