+++ analysis/auxil/parameter_derivation.py
