+++ observations/auxil/read_lightcurve.py
