+++ observations/auxil/read_spectrum.py
